feat: Allow setting custom precision for printing profiling data
