+++ cmd/run.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"time"
 
 	"buf.build/gen/go/cedana/cedana/protocolbuffers/go/daemon"
 	"github.com/cedana/cedana/pkg/client"
@@ -12,11 +11,7 @@ import (
 	"github.com/cedana/cedana/pkg/features"
 	"github.com/cedana/cedana/pkg/flags"
 	"github.com/cedana/cedana/pkg/keys"
-	"github.com/cedana/cedana/pkg/profiling"
-	"github.com/cedana/cedana/pkg/style"
 	"github.com/cedana/cedana/pkg/utils"
-	"github.com/jedib0t/go-pretty/v6/table"
-	"github.com/jedib0t/go-pretty/v6/text"
 	"github.com/spf13/cobra"
 )
 
@@ -177,65 +172,4 @@ var processRunCmd = &cobra.Command{
 
 		return nil
 	},
-}
-
-////////////////////
-/// Helper Funcs ///
-////////////////////
-
-// PrintProfilingData prints the profiling data in a very readable format.
-func printProfilingData(data *profiling.Data) {
-	var total time.Duration
-
-	fmt.Print("Profiling data received from daemon:\n\n")
-
-	tableWriter := table.NewWriter()
-	tableWriter.SetStyle(style.TableStyle)
-	tableWriter.SetOutputMirror(os.Stdout)
-
-	categoryMap := make(map[string]time.Duration)
-
-	for _, p := range data.Components {
-		if p.Duration == 0 {
-			continue
-		}
-		categoryName, name := utils.SimplifyFuncName(p.Name)
-
-		category := style.WarningColors.Sprint(categoryName)
-		features.CmdTheme.IfAvailable(func(name string, theme text.Colors) error {
-			category = theme.Sprint(categoryName)
-			return nil
-		}, categoryName)
-
-		duration := time.Duration(p.Duration)
-		total += duration
-
-		if categoryName != "" {
-			categoryMap[category] += duration
-		} else {
-			categoryMap[style.DisabledColors.Sprint("other")] += duration
-		}
-
-		tableWriter.AppendRow([]interface{}{duration, category, style.DisabledColors.Sprint(name)})
-	}
-
-	tableWriter.AppendFooter([]interface{}{total, "", fmt.Sprintf("%s (total)", data.Name)})
-	tableWriter.Render()
-
-	if len(categoryMap) > 1 {
-		fmt.Println()
-		tableWriter = table.NewWriter()
-		tableWriter.SetStyle(style.TableStyle)
-		tableWriter.SetOutputMirror(os.Stdout)
-
-		for category, duration := range categoryMap {
-			percentage := (float64(duration) / float64(total)) * 100
-			tableWriter.AppendRow([]interface{}{duration, fmt.Sprintf("%.2f%%", percentage), category})
-		}
-
-		tableWriter.AppendFooter([]interface{}{total, "", fmt.Sprintf("%s (total)", data.Name)})
-		tableWriter.Render()
-	}
-
-	fmt.Println()
 }
