+++ pkg/config/config.go
@@ -48,7 +48,8 @@ var Global Config = Config{
 		Path:   filepath.Join(os.TempDir(), "cedana.db"),
 	},
 	Profiling: Profiling{
-		Enabled: true,
+		Enabled:       true,
+		Precision: "auto",
 	},
 	Metrics: Metrics{
 		ASR:  false,
